fix some type warnings
(cherry picked from commit 84b20b4363aa205b775f5ca1b2084aa70c4c0440)

+++ distributedcombigrid/tests/TaskCount.hpp
@@ -43,7 +43,7 @@ class TaskCount : public combigrid::Task {
       p = {nprocs,1};
     } else {
       for (const auto& d : decomposition) {
-        p.push_back(d.size());
+        p.push_back(static_cast<int>(d.size()));
       }
     }
     dfg_ = new DistributedFullGrid<CombiDataType>(getDim(), getLevelVector(), lcomm, getBoundary(), p, false, decomposition);
